add support for immutable collections
