+++ gradle/libs.versions.toml
@@ -14,7 +14,7 @@ mokoResourcesVersion = "0.16.2"
 kotlinxMetadataKLibVersion = "0.0.1"
 
 kotlinPoetVersion = "1.6.0"
-swiftPoetVersion = "1.5.1.3"
+swiftPoetVersion = "1.5.1.4"
 
 [libraries]
 appCompat = { module = "androidx.appcompat:appcompat", version.ref = "androidAppCompatVersion" }
